Separated generic begin_parallel_x events and tuning from backend-accessible functionality
