+++ core/src/impl/Kokkos_ViewMapping.hpp
@@ -2930,8 +2930,9 @@ struct ViewValueFunctor<DeviceType, ValueType, false /* is_scalar */> {
         functor_name =
             (destroy ? "Kokkos::View::destruction [" + name + "]"
                      : "Kokkos::View::initialization [" + name + "]");
-        Kokkos::Tools::Impl::begin_parallel_for(policy, *this, functor_name,
-                                                kpID);
+        Kokkos::Profiling::beginParallelFor(
+            "Kokkos::View::initialization [" + name + "]",
+            Kokkos::Profiling::Experimental::device_id(space), &kpID);
       }
 
 #ifdef KOKKOS_ENABLE_CUDA
@@ -2945,8 +2946,7 @@ struct ViewValueFunctor<DeviceType, ValueType, false /* is_scalar */> {
       closure.execute();
       space.fence("Kokkos::Impl::ViewValueFunctor: View init/destroy fence");
       if (Kokkos::Profiling::profileLibraryLoaded()) {
-        Kokkos::Tools::Impl::end_parallel_for(policy, *this, functor_name,
-                                              kpID);
+        Kokkos::Profiling::endParallelFor(kpID);
       }
     } else {
       for (size_t i = 0; i < n; ++i) operator()(i);
@@ -3025,8 +3025,9 @@ struct ViewValueFunctor<DeviceType, ValueType, true /* is_scalar */> {
       std::string functor_name = "Kokkos::View::initialization [" + name + "]";
       uint64_t kpID            = 0;
       if (Kokkos::Profiling::profileLibraryLoaded()) {
-        Kokkos::Tools::Impl::begin_parallel_for(policy, *this, functor_name,
-                                                kpID);
+        Kokkos::Profiling::beginParallelFor(
+            "Kokkos::View::initialization [" + name + "]",
+            Kokkos::Profiling::Experimental::device_id(space), &kpID);
       }
 #ifdef KOKKOS_ENABLE_CUDA
       if (std::is_same<ExecSpace, Kokkos::Cuda>::value) {
@@ -3040,8 +3041,7 @@ struct ViewValueFunctor<DeviceType, ValueType, true /* is_scalar */> {
       space.fence(
           "Kokkos::Impl::ViewValueFunctor: Fence after setting values in view");
       if (Kokkos::Profiling::profileLibraryLoaded()) {
-        Kokkos::Tools::Impl::end_parallel_for(policy, *this, functor_name,
-                                              kpID);
+        Kokkos::Profiling::endParallelFor(kpID);
       }
     } else {
       for (size_t i = 0; i < n; ++i) operator()(i);
